+++ tests/mock_host_lib/__init__.py
@@ -1,4 +1,4 @@
 """Exposed mock_host_lib classes and methods."""
-from tests.mock_host_lib.getters import mock_os_release, mock_os_release_tuple
+from tests.mock_host_lib.getters import mock_os_release, mock_os_release_tuple, mock_uname_tuple, mock_uname
 
 mock_command_executor_method = "shell_executor_lib.CommandManager.execute_command"