+++ tests/mock_host_lib/getters/__init__.py
@@ -1,2 +1,3 @@
 """Exposed getters classes and methods."""
 from tests.mock_host_lib.getters.mock_os_getter import mock_os_release, mock_os_release_tuple
+from tests.mock_host_lib.getters.mock_uname_getter import mock_uname_tuple, mock_uname