+++ tests/mock_host_lib/getters/mock_uname_getter.py
@@ -1,0 +1,5 @@
+"""Mocks of uname getter."""
+
+mock_uname = ['debian 6.1.0-9-amd64']
+
+mock_uname_tuple = ("debian", "6.1.0-9-amd64")